+++ src/App.js
@@ -1,13 +1,15 @@
 import React, { Component, Suspense } from 'react';
 import './App.css';
-import L18nextComponent from "./components/I18next";
+import NewsFeed from "./components/NewsFeed"
+import Header from "./components/Header"
 
 
 class App extends React.Component {
   render() {
     return (
       <div className="App">
-        <L18nextComponent/>
+        <Header />
+        <NewsFeed/>
       </div>
     );
   }