+++ src/i18n.js
@@ -21,7 +21,9 @@ i18n
 .init({
 	fallbackLng: 'en',
 	debug: true,
-
+	react: {
+		useSuspense: false //   <---- this will do the magic
+	},
 	interpolation: {
 		escapeValue: false, // not needed for react as it escapes by default
 	}