+++ src/support/APIkeys/index.js
@@ -1,0 +1,3 @@
+export default {
+	googleNews: '27b3dfca16e7435c8b9ca9fbbb0d1edc'
+};