+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0
+* ADD: Google News API, fetch news 
+* ADD: Header, HewsFeed components
+* FIX: Set i18 to not use Suspense to avoid stupid errors  
+
 ## 0.2.0
 * ADD: husky, fetch-mock, jest